Refactoring part II
